## 1.0.15-beta (Jan 10, 2023) BUGFIXES: * `Offset` and `Limit` query params change to `int`.

+++ dnacenter/resource_applications.go
@@ -1273,8 +1273,8 @@ func searchApplicationPolicyGetApplications(m interface{}, queryParams dnacenter
 					return foundItem, err
 				}
 			}
-			queryParams.Limit = float64(maxPageSize)
-			queryParams.Offset += float64(maxPageSize)
+			queryParams.Limit = maxPageSize
+			queryParams.Offset += maxPageSize
 			nResponse, _, err = client.ApplicationPolicy.GetApplications(&queryParams)
 		}
 		return nil, err
